Added scalar multiplication of vectors for primative number types
This feature allows the programmer to multiply a vector by a scalar
number on the left hand side. If the type is 'T' and the vector is 'V',
then you can now do T * V to scale the vector. This feature is included for
f64, f32, i8, i16, i32, i64, i128, isize, u8, u16, u32, u64, u128, and usize.